requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.0
sqlalchemy==2.0.25
vanna==0.5.5
python-dotenv==1.0.0
